Refactored _updateObject function into new actor helper class Updated function to handle new 7.1 formData structure.
